requirements.txt
discord.py>=2.3.0
requests>=2.28.0
asyncio